For #314 Removed vector virtual T_sp& operator[]
It was a mistake to implement an operator that returns a T_sp reference
to elements of vectors because specialized vectors won't contain T_sp
objects.   Use elt and setf_elt instead because they can convert
arguments and return values

+++ include/clasp/core/array.h
@@ -58,7 +58,7 @@ public:
 
 public: // Functions here
   virtual bool equalp(T_sp other) const { SUBIMP(); };
-  virtual T_sp aset_unsafe(int j, T_sp val) { SUBIMP(); };
+  virtual T_sp aset_unsafe(size_t j, T_sp val) { SUBIMP(); };
 CL_LISPIFY_NAME("cl:arrayHasFillPointerP");
 CL_DEFMETHOD   virtual bool arrayHasFillPointerP() const { return false; };
   virtual gc::Fixnum arrayTotalSize() const;
