+++ src/core/str.cc
@@ -197,7 +197,7 @@ T_sp Str_O::setf_elt(int index, T_sp val) {
   return val;
 }
 
-T_sp Str_O::aset_unsafe(int index, T_sp val) {
+T_sp Str_O::aset_unsafe(size_t index, T_sp val) {
   _OF();
   char ch = clasp_as_char(gc::As<Character_sp>(val));
   this->_Contents[index] = ch;
